move some event handling around
